Improved application of styles
Use <span> instead of <p> for added styles so that applying a new style does not introduce a new block.

Replaces the default FormatBlock algorithm used when a style is applied with a new method that treats styles based on inline elements from those based on block based elements. Custom styles based on inline tags are applied using the same method as applying fonts. Custom styles based on block tags replace the block tags in any selected HTML with the tag from the style splitting nodes as needed.